Defer importing any NeMo-related code until block's start() method
gnuradio-companion creates embedded python blocks via block.__init__() to check them, in the
main GUI process, before generating and calling the subprocess that will run the flowgraph.
This means that anything that allocates significant amounts of memory must be deferred to
start() or it will be allocated twice -- and both allocations would be kept around for the
entire duration of the run.